refactor: rename common dir to compiler
