+++ src/compiler/dagFactory.ts
